fix: fix rounding error by casting for marts

+++ medicare_dbt/models/marts/mart_enrollment_by_state.sql
@@ -14,8 +14,8 @@ by_state as (
         disabled_beneficiaries,
         male_beneficiaries,
         female_beneficiaries,
-        ROUND(medicare_advantage_benes / NULLIF(total_beneficiaries, 0) * 100, 1) AS ma_penetration_rate,
-        ROUND(original_medicare_benes / NULLIF(total_beneficiaries, 0) * 100, 1)  AS ffs_penetration_rate        
+        ROUND((medicare_advantage_benes / NULLIF(total_beneficiaries, 0) * 100)::numeric, 1) AS ma_penetration_rate,
+        ROUND((original_medicare_benes / NULLIF(total_beneficiaries, 0) * 100)::numeric, 1)  AS ffs_penetration_rate,
     from staging
     where bene_geo_lvl = 'State'
     and month = 'Year'
